Update Per Diem API Automation.py
This should work for general use! You will need to get an API key from the GSA and I will work on updating the final pop-up to download an Excel file with the numbers but this will provide a quick and easy solution. If you are having issues, comment out everything from below the imported packages until the API call and enter the dates in the code.